requirements.txt
ultralytics==8.1.0
opencv-python==4.9.0.80
albumentations==1.3.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
PyYAML==6.0.1
numpy==1.26.4
torch==2.1.0  # Более стабильная версия PyTorch
torchvision==0.16.0